requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Scheduling
apscheduler==3.11.0

# Google Sheets
gspread==6.1.4
google-auth==2.38.0

# AI
anthropic==0.44.0

# Email
resend==2.7.0

# Utilities
python-dotenv==1.0.1
pydantic==2.11.1
pydantic-settings==2.8.1
httpx==0.28.1
python-dateutil==2.9.0
